Create Automate_linkden_connection.py
LinkedIn connections are a very important thing for an IT professional, so we need to send connection requests to a lot of people who can be useful to us. But sometimes sending connection requests one at a time can be a little annoying and hectic. It would be nice to automate this work but How? 
Python to rescue!
In this article, we will learn how to automate the accepting of LinkedIn connections using Python.
Modules required – 
 

Selenium – Selenium does not comes built-in with python. To install selenium type the below command in the terminal. 
 
pip install selenium
 
Pyautogui – Pyautogui also does not comes built-in with python. To install pyautogui type the below command in the terminal. 
 
pip install pyautogui
 
Chrome web driver –  download chrome web driver